+++ awint_macro_internals/src/lib.rs
@@ -48,6 +48,7 @@ pub fn awint_macro_cc(input: &str) -> Result<String, String> {
         static_construction_fn: awint_static_construction_fn,
         lit_construction_fn: awint_unreachable_construction_fn,
         construction_fn: cc_construction_fn,
+        const_wrapper: identity_const_wrapper,
         fn_names: AWINT_FN_NAMES,
     };
     cc_macro(input, code_gen, AWINT_NAMES)
@@ -61,6 +62,7 @@ pub fn awint_macro_inlawi(input: &str) -> Result<String, String> {
         static_construction_fn: awint_static_construction_fn,
         lit_construction_fn: awint_inlawi_lit_construction_fn,
         construction_fn: inlawi_construction_fn,
+        const_wrapper: identity_const_wrapper,
         fn_names: AWINT_FN_NAMES,
     };
     cc_macro(input, code_gen, AWINT_NAMES)
@@ -74,6 +76,7 @@ pub fn awint_macro_extawi(input: &str) -> Result<String, String> {
         static_construction_fn: awint_static_construction_fn,
         lit_construction_fn: awint_extawi_lit_construction_fn,
         construction_fn: extawi_construction_fn,
+        const_wrapper: identity_const_wrapper,
         fn_names: AWINT_FN_NAMES,
     };
     cc_macro(input, code_gen, AWINT_NAMES)
@@ -87,6 +90,7 @@ pub fn awint_macro_bits(input: &str) -> Result<String, String> {
         static_construction_fn: awint_static_construction_fn,
         lit_construction_fn: awint_bits_lit_construction_fn,
         construction_fn: inlawi_construction_fn,
+        const_wrapper: awint_bits_const_wrapper,
         fn_names: AWINT_FN_NAMES,
     };
     cc_macro(input, code_gen, AWINT_NAMES)